+++ src/Module_Particles/microcode_generator.h
@@ -2,8 +2,8 @@
 
 namespace Lab {
 
-auto buffer_to_vec(size_t dimension, std::string const& buffer_name, std::string const& id_name) -> std::string;
-auto vec_to_buffer(size_t dimension, std::string const& buffer_name, std::string const& id_name, std::string const& vec_name) -> std::string;
-auto vec_zero(size_t dimension) -> std::string;
+auto buffer_to_vec(int dimension, std::string const& buffer_name, std::string const& id_name) -> std::string;
+auto vec_to_buffer(int dimension, std::string const& buffer_name, std::string const& id_name, std::string const& vec_name) -> std::string;
+auto vec_zero(int dimension) -> std::string;
 
 }; // namespace Lab
